requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
matplotlib
python-dotenv
sqlalchemy
